remove buttons and buggy zoom
Button added to ‘remove bad’ and ‘remove ok’ vowels.  Select which
removal option, hold ctrl and click vowel to remove it without a
comment. Hold ctrl and shift to remove with a comment.

First attempts at zooming work some of the time but are buggy.  TODO:
define a vowel plot class that automatically resizes when max-min are
changed (this should deal with current problems)

Vowels are now a class with their own function to make an alternate
pronunciation.