start_plotmish.py + updateFormants.py
Includes start_plotmish.py which is a command line gui for plotmish.
For the moment it only works if it is in a folder with the main
plotmish folder (move it before you can use it).

updateFormants.py takes the existing log files and rewrites the old
formant.txt files